commitchanges

+++ APITaskCore/Migrations/ApplicationContextModelSnapshot.cs
@@ -21,9 +21,9 @@ namespace APITaskCore.Migrations
 
             modelBuilder.Entity("APITaskCore.Models.Breweries", b =>
                 {
-                    b.Property<long>("BrewId")
+                    b.Property<int>("BrewId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<double>("AverageRating")
@@ -86,8 +86,8 @@ namespace APITaskCore.Migrations
                     b.Property<int>("BrewId")
                         .HasColumnType("int");
 
-                    b.Property<long?>("BreweryBrewId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("BreweryBrewId")
+                        .HasColumnType("int");
 
                     b.Property<string>("CreatedBy")
                         .HasColumnType("nvarchar(max)");
